+++ tools/dataset_converters/indoor_converter.py
@@ -4,9 +4,10 @@ import os
 import mmcv
 import numpy as np
 
-from tools.data_converter.s3dis_data_utils import S3DISData, S3DISSegData
-from tools.data_converter.scannet_data_utils import ScanNetData, ScanNetSegData
-from tools.data_converter.sunrgbd_data_utils import SUNRGBDData
+from tools.dataset_converters.s3dis_data_utils import S3DISData, S3DISSegData
+from tools.dataset_converters.scannet_data_utils import (ScanNetData,
+                                                         ScanNetSegData)
+from tools.dataset_converters.sunrgbd_data_utils import SUNRGBDData
 
 
 def create_indoor_info_file(data_path,
